Dynamic status pages on graphic display
Allows to display graphic status pages and browse them with rotary
encoder.

I developped this library to run it on a Teensy 3.2 ARM board but if
should be compatible with Arduino. The display I use is a OLED SSD 1322
(256x64 pixels) with grey shade, but you can change that and take
advantage of all the displays supported by U8Glib.

This library use U8Glib.h and ClickEncoder.h in order to work, they are
included in a separate ZIP file.